\name{DKAP1061}
\encoding{UTF-8}
\alias{DKAP1061}
\title{DKAP1061 Dataset}
\description{
DKAP1061 is a dataset from a survey on students' digital competence, including demographic and educational background variables.
}
\usage{
data(DKAP1061)
}

\format{
A data frame with multiple columns. Selected variables:

\describe{
  \item{\code{ecostt}}{Student's family economic status.}
  \item{\code{edufat}}{Father's education level.}
  \item{\code{edumot}}{Mother's education level.}
  \item{\code{ict}}{Digital competence score.}
  \item{\code{mean_dr}}{Mean digital resources.}
  \item{\code{mean_ict}}{Mean ICT skills score.}
  \item{\code{mean_il}}{Mean information literacy score.}
  \item{\code{mean_ppr}}{Mean personal productivity rating.}
  \item{\code{mean_udcr}}{Mean use of digital content/resources.}
  \item{\code{schoolid}}{School ID.}
  \item{\code{schid}}{School code (alternative to schoolid).}
  \item{\code{sex}}{Student's gender (1 = female, 2 = male).}
  \item{\code{stuid}}{Student ID.}

  \item{\code{a1}}{Not yet documented.}
  \item{\code{a10}}{Not yet documented.}
  \item{\code{a11}}{Not yet documented.}
  \item{\code{a12}}{Not yet documented.}
  \item{\code{a13}}{Not yet documented.}
  \item{\code{a14}}{Not yet documented.}
  \item{\code{a2}}{Not yet documented.}
  \item{\code{a3}}{Not yet documented.}
  \item{\code{a4}}{Not yet documented.}
  \item{\code{a5}}{Not yet documented.}
  \item{\code{a6}}{Not yet documented.}
  \item{\code{a7}}{Not yet documented.}
  \item{\code{a8}}{Not yet documented.}
  \item{\code{a9}}{Not yet documented.}
  \item{\code{b1}}{Not yet documented.}
  \item{\code{b10}}{Not yet documented.}
  \item{\code{b11}}{Not yet documented.}
  \item{\code{b12}}{Not yet documented.}
  \item{\code{b13}}{Not yet documented.}
  \item{\code{b14}}{Not yet documented.}
  \item{\code{b15_1}}{Not yet documented.}
  \item{\code{b15_2}}{Not yet documented.}
  \item{\code{b15_3}}{Not yet documented.}
  \item{\code{b15_4}}{Not yet documented.}
  \item{\code{b15_5}}{Not yet documented.}
  \item{\code{b15_6}}{Not yet documented.}
  \item{\code{b15_7}}{Not yet documented.}
  \item{\code{b15_8}}{Not yet documented.}
  \item{\code{b16_1}}{Not yet documented.}
  \item{\code{b16_2}}{Not yet documented.}
  \item{\code{b16_3}}{Not yet documented.}
  \item{\code{b16_4}}{Not yet documented.}
  \item{\code{b16_5}}{Not yet documented.}
  \item{\code{b16_6}}{Not yet documented.}
  \item{\code{b16_7}}{Not yet documented.}
  \item{\code{b16_8}}{Not yet documented.}
  \item{\code{b17_1}}{Not yet documented.}
  \item{\code{b17_2}}{Not yet documented.}
  \item{\code{b17_3}}{Not yet documented.}
  \item{\code{b17_4}}{Not yet documented.}
  \item{\code{b17_5}}{Not yet documented.}
  \item{\code{b17_6}}{Not yet documented.}
  \item{\code{b17_7}}{Not yet documented.}
  \item{\code{b17_8}}{Not yet documented.}
  \item{\code{b18_1}}{Not yet documented.}
  \item{\code{b18_2}}{Not yet documented.}
  \item{\code{b18_3}}{Not yet documented.}
  \item{\code{b18_4}}{Not yet documented.}
  \item{\code{b18_5}}{Not yet documented.}
  \item{\code{b18_6}}{Not yet documented.}
  \item{\code{b18_7}}{Not yet documented.}
  \item{\code{b18_8}}{Not yet documented.}
  \item{\code{b18_9}}{Not yet documented.}
  \item{\code{b2}}{Not yet documented.}
  \item{\code{b3}}{Not yet documented.}
  \item{\code{b4}}{Not yet documented.}
  \item{\code{b5}}{Not yet documented.}
  \item{\code{b6}}{Not yet documented.}
  \item{\code{b7}}{Not yet documented.}
  \item{\code{b8}}{Not yet documented.}
  \item{\code{b9}}{Not yet documented.}
  \item{\code{c1}}{Not yet documented.}
  \item{\code{c10}}{Not yet documented.}
  \item{\code{c11}}{Not yet documented.}
  \item{\code{c12}}{Not yet documented.}
  \item{\code{c2}}{Not yet documented.}
  \item{\code{c3}}{Not yet documented.}
  \item{\code{c4}}{Not yet documented.}
  \item{\code{c5}}{Not yet documented.}
  \item{\code{c6}}{Not yet documented.}
  \item{\code{c7}}{Not yet documented.}
  \item{\code{c8}}{Not yet documented.}
  \item{\code{c9}}{Not yet documented.}
  \item{\code{d1}}{Not yet documented.}
  \item{\code{d10}}{Not yet documented.}
  \item{\code{d11}}{Not yet documented.}
  \item{\code{d12}}{Not yet documented.}
  \item{\code{d13}}{Not yet documented.}
  \item{\code{d14}}{Not yet documented.}
  \item{\code{d15}}{Not yet documented.}
  \item{\code{d16}}{Not yet documented.}
  \item{\code{d2}}{Not yet documented.}
  \item{\code{d3}}{Not yet documented.}
  \item{\code{d4}}{Not yet documented.}
  \item{\code{d5}}{Not yet documented.}
  \item{\code{d6}}{Not yet documented.}
  \item{\code{d7}}{Not yet documented.}
  \item{\code{d8}}{Not yet documented.}
  \item{\code{d9}}{Not yet documented.}
  \item{\code{e1}}{Not yet documented.}
  \item{\code{e10}}{Not yet documented.}
  \item{\code{e11}}{Not yet documented.}
  \item{\code{e2}}{Not yet documented.}
  \item{\code{e3}}{Not yet documented.}
  \item{\code{e4}}{Not yet documented.}
  \item{\code{e5}}{Not yet documented.}
  \item{\code{e6}}{Not yet documented.}
  \item{\code{e7}}{Not yet documented.}
  \item{\code{e8}}{Not yet documented.}
  \item{\code{e9}}{Not yet documented.}
  \item{\code{f1}}{Not yet documented.}
  \item{\code{f2}}{Not yet documented.}
  \item{\code{f3}}{Not yet documented.}
  \item{\code{f4}}{Not yet documented.}
  \item{\code{f5}}{Not yet documented.}
  \item{\code{f6}}{Not yet documented.}
  \item{\code{f7}}{Not yet documented.}
  \item{\code{f8_1}}{Not yet documented.}
  \item{\code{f8_2}}{Not yet documented.}
  \item{\code{f8_3}}{Not yet documented.}
  \item{\code{f8_4}}{Not yet documented.}
  \item{\code{f8_5}}{Not yet documented.}
  \item{\code{g1}}{Not yet documented.}
  \item{\code{g10_1}}{Not yet documented.}
  \item{\code{g10_2}}{Not yet documented.}
  \item{\code{g10_3}}{Not yet documented.}
  \item{\code{g11}}{Not yet documented.}
  \item{\code{g12}}{Not yet documented.}
  \item{\code{g13}}{Not yet documented.}
  \item{\code{g14}}{Not yet documented.}
  \item{\code{g15}}{Not yet documented.}
  \item{\code{g16}}{Not yet documented.}
  \item{\code{g17}}{Not yet documented.}
  \item{\code{g18}}{Not yet documented.}
  \item{\code{g2}}{Not yet documented.}
  \item{\code{g3_1}}{Not yet documented.}
  \item{\code{g3_2}}{Not yet documented.}
  \item{\code{g3_3}}{Not yet documented.}
  \item{\code{g3_4}}{Not yet documented.}
  \item{\code{g4_1}}{Not yet documented.}
  \item{\code{g4_2}}{Not yet documented.}
  \item{\code{g4_3}}{Not yet documented.}
  \item{\code{g4_4}}{Not yet documented.}
  \item{\code{g4_5}}{Not yet documented.}
  \item{\code{g4_6}}{Not yet documented.}
  \item{\code{g5_1}}{Not yet documented.}
  \item{\code{g5_2}}{Not yet documented.}
  \item{\code{g5_3}}{Not yet documented.}
  \item{\code{g5_4}}{Not yet documented.}
  \item{\code{g5_5}}{Not yet documented.}
  \item{\code{g5_6}}{Not yet documented.}
  \item{\code{g6_1}}{Not yet documented.}
  \item{\code{g6_2}}{Not yet documented.}
  \item{\code{g6_3}}{Not yet documented.}
  \item{\code{g6_4}}{Not yet documented.}
  \item{\code{g6_5}}{Not yet documented.}
  \item{\code{g6_6}}{Not yet documented.}
  \item{\code{g7_1}}{Not yet documented.}
  \item{\code{g7_2}}{Not yet documented.}
  \item{\code{g7_3}}{Not yet documented.}
  \item{\code{g8_1}}{Not yet documented.}
  \item{\code{g8_2}}{Not yet documented.}
  \item{\code{g8_3}}{Not yet documented.}
  \item{\code{g9}}{Not yet documented.}
  \item{\code{h1_1}}{Not yet documented.}
  \item{\code{h1_2}}{Not yet documented.}
  \item{\code{h1_3}}{Not yet documented.}
  \item{\code{h1_4}}{Not yet documented.}
  \item{\code{h1_5}}{Not yet documented.}
  \item{\code{h1_6}}{Not yet documented.}
  \item{\code{h1_7}}{Not yet documented.}
  \item{\code{h2}}{Not yet documented.}
  \item{\code{h3}}{Not yet documented.}
  \item{\code{h4_1}}{Not yet documented.}
  \item{\code{h4_2}}{Not yet documented.}
  \item{\code{h4_3}}{Not yet documented.}
  \item{\code{h5}}{Not yet documented.}
  \item{\code{h6_1}}{Not yet documented.}
  \item{\code{h6_2}}{Not yet documented.}
  \item{\code{h6_3}}{Not yet documented.}
  \item{\code{h6_4}}{Not yet documented.}
  \item{\code{h7_1}}{Not yet documented.}
  \item{\code{h7_2}}{Not yet documented.}
  \item{\code{h7_3}}{Not yet documented.}
  \item{\code{h7_4}}{Not yet documented.}

}
\emph{Note: Variables starting with \code{a}, \code{b}, \code{c}, \code{d}, \code{f}, \code{g}, \code{h} are omitted from this documentation.}
}

\references{
For documentation, case studies, and examples, visit the GitHub repository:

\itemize{
  \item \href{https://github.com/sshpa/bayesvl/tree/master/References}{Documentation and User Guide}
  \item \href{https://www.nature.com/articles/s41599-020-0442-3}{Published Case Study (Vuong & La, 2020)}
}
}
\examples{
data(DKAP1061)

# Preview the dataset
head(DKAP1061)
}
\keyword{datasets}
