\name{Legends345}
\alias{Legends345}
\title{Legends345 data}
\description{
  Legends345.
}
\usage{
	data(Legends345)
}
\format{
    \enumerate{
        \item O : Whether or not happy ending for main character
        \item VB : Whether or not the main character behaves in accordance with the core values of Buddhism
        \item VC : Whether or not the main character behaves in accordance with the core values of Confucianism
        \item VT : Whether or not the main character behaves in accordance with the core values of Taoism
        \item Lie : Whether or not the main character tells lie
        \item Viol : Whether or not the main character commits acts of violence
        \item Int1 : Whether there are interventions from the supernatural world
        \item Int2 : Whether there are interventions from the human world
    }
}
\references{
For documentation, case studies, worked examples, and other tutorial materials, visit the References section on our GitHub:

\itemize{
  \item \href{https://github.com/sshpa/bayesvl/tree/master/References}{https://github.com/sshpa/bayesvl/tree/master/References}
}

For case studies using the package in research articles, see:

\itemize{
  \item \href{https://www.nature.com/articles/s41599-020-0442-3}{https://www.nature.com/articles/s41599-020-0442-3}
}
}
\examples{
	data(Legends345)
	
	data1 <- Legends345
	head(data1)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{legends}

