\name{Legends345}
\alias{Legends345}
\title{Legends345 data}
\description{
  Legends345.
}
\usage{
	data(Legends345)
}
\format{
    \enumerate{
        \item O : Number of fish caught during visit
        \item VB : Whether or not group used livebait to fish
        \item VC : Whether or not group had a camper
        \item VT : Number of adults in group
        \item Lie : Number of children in group
        \item Viol : Number of hours group spent in park
    }
}
\references{
	reserved for manual link
}
\examples{
	data(Legends345)
	
	data1 <- Legends345
	head(data1)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{legends}

